[ICC-23] Service 코드 수정 및 생성

+++ src/main/java/com/icc/qasker/quiz/dto/response/ResultResponse.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class ResultResponse {
-    private Long number;
+    private int number;
     private Long correctAnswer;
     private boolean isCorrect;
     private String explanation;
